1차 완성 (CRUD, 회원가입, 로그인 병합)

+++ src/main/java/com/sparta/team9back/service/PostService.java
@@ -3,6 +3,7 @@ package com.sparta.team9back.service;
 import com.sparta.team9back.dto.PostRequestDto;
 import com.sparta.team9back.dto.PostResponseDto;
 import com.sparta.team9back.model.Post;
+import com.sparta.team9back.model.User;
 import com.sparta.team9back.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,9 +16,10 @@ public class PostService {
     private final PostRepository postRepository;
 
     @Transactional // 이걸 잊지 않았나 하는 생각.
-    public PostResponseDto createPost(PostRequestDto postRequestDto) {
+    public PostResponseDto createPost(PostRequestDto postRequestDto, User user) {
 
         Post post = Post.builder()
+                .user(user)
                 .goodsImg(postRequestDto.getGoodsImg())
                 .title(postRequestDto.getTitle())
                 .content(postRequestDto.getContent())
@@ -28,6 +30,7 @@ public class PostService {
         postRepository.save(post);
 
         return PostResponseDto.builder()
+                .username(user.getUsername())
                 .title(postRequestDto.getTitle())
                 .content(postRequestDto.getContent())
                 .category(postRequestDto.getCategory())
@@ -38,7 +41,7 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponseDto showDetail(Long postId) {
+    public PostResponseDto showDetail(Long postId, User user) {
 
         Post post = postRepository.findById(postId).orElse(null);
         if (post == null) {
@@ -46,6 +49,7 @@ public class PostService {
         }
 
         return PostResponseDto.builder()
+                .username(user.getUsername())
                 .title(post.getTitle())
                 .content(post.getContent())
                 .category(post.getCategory())
@@ -56,19 +60,19 @@ public class PostService {
     }
 
     @Transactional
-    public Long updatePost(Long postId, PostRequestDto postRequestDto) {
-        Post post = postRepository.findById(postId).orElseThrow(
-                () -> new IllegalArgumentException("해당 게시글 아이디가 존재하지 않습니다.")
+    public void updatePost(Long postId, PostRequestDto postRequestDto, User user) {
+        Post post = postRepository.findByUserAndPostId(user, postId).orElseThrow(
+                () -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다.")
         );
         post.update(postRequestDto);
-        return post.getPostId();
     }
 
     @Transactional
-    public Long deletePost(Long postId) {
+    public void deletePost(Long postId, User user) {
+        Long checkId = postRepository.findByUser(user).orElse(null).getPostId();
+        if(checkId.equals(postId))
+            postRepository.deleteById(postId);
         // 댓글을 추가할 경우, 그리고 Post와 Comment entity 간에 cascade 설정을 해놓지 않은 경우
         // 댓글이 달린 게시글을 삭제할 때 Service에서 게시글에 달린 댓글을 전부 삭제하는 작업이 선행될 필요가 있음
-        postRepository.deleteById(postId);
-        return postId;
     }
 }
